Move loader and entryPoints into esbuild-options module
